# Services/AiVisionService/openapi.json
{"openapi": "3.0.2", "info": {"title": "Object Detection FastAPI Template", "description": "Obtain object value out of image\n                    and return image and json result", "version": "2023.1.31"}, "paths": {"/healthcheck": {"get": {"summary": "Perform Healthcheck", "description": "It basically sends a GET request to the route & hopes to get a \"200\"\nresponse code. Failing to return a 200 response code just enables\nthe GitHub Actions to rollback to the last version the project was\nfound in a \"working condition\". It acts as a last line of defense in\ncase something goes south.\nAdditionally, it also returns a JSON response in the form of:\n{\n    'healtcheck': 'Everything OK!'\n}", "operationId": "perform_healthcheck_healthcheck_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/vision/segment": {"post": {"summary": "Segment Image", "description": "Object Detection from an image URL.\nReturns:\n    dict: JSON format matching AiSegmentationResultDto", "operationId": "segment_image_api_v1_vision_segment_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SegmentationRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/vision/segment/draw": {"post": {"summary": "Segment Image Draw", "description": "Object Detection from an image URL and returns drawn image.\nReturns:\n    StreamingResponse: JPEG image binary", "operationId": "segment_image_draw_api_v1_vision_segment_draw_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SegmentationRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/vision/colorize": {"post": {"summary": "Colorize Image", "description": "Object Colorization from an image URL using U-Net.\nReturns:\n    StreamingResponse: PNG image binary", "operationId": "colorize_image_api_v1_vision_colorize_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SegmentationRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "SegmentationRequest": {"title": "SegmentationRequest", "required": ["imageUrl"], "type": "object", "properties": {"imageUrl": {"title": "Imageurl", "type": "string"}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}